Modified master metadata to not keep track of pages
As discussed, we won't necessarily need the master to know which worker
stores which pages, since most of the time there is one worker that
stores a block. The new metadata will only store, with each page, a list
of storage dir ids that parts of the block are stored on.

This requires undoing some of the metadata changes we made and then
allowing the master to store multiple storage dirs per worker (it
originally stores only one). We also modified the RemoteBlockInStream
class to take advantage of this storage directory information (it
originally didn't know anything about storage directories) and
prioritize reading from higher storage tiers.