Modify worker to store blocks as pages
The core change here is to BlockHandlerLocal. Now the block location is
a directory which stores numerous page files inside. The BlockHandler
should almost always be used to read and write blocks. We modified the
BlockHandler interface slightly, to allow appending only to the end of a
block (not anywhere), fetching channels for a specific range of the
block (not the entire block), and a few others. Also we added buffering
functionality to BlockHandlerLocal (so that numerous small writes don't
cost too much)

Also we modified BlockOutStream to use a BlockHandler to write and
TachyonFile to use a BlockHandler to read a block. Most of the other
changes involve updating Tachyon to deal with blocks as directories (and
use BlockHandler where appropriate) rather than files.