Changed getChannel to getChannels
Now the BlockHandler can return a list of channels that cover the given
offset-length range. This method can now serve as the basis for the read
method in BlockHandlerLocal as well. We also added more tests for this
in BlockHandlerLocal. This change simplifies the BlockResponse encode
method for the TRANSFER file transfer type.