Implemented cacheBlock as a block move
We're preparing for the case where two people want to concurrently cache
different partially-cached editions of a block. In this case, we want to
move all the pages from both user temp directories to the shared
destination directory, so simply renaming one of the directories and
then failing the other would be suboptimal.

This is sometimes less performant than renaming the directory, but that
can be worked on, and there's a comment addressing that in
BlockHandlerLocal.