Split BlockHandlerLocal into three different ones
Rather than combining all the functionality of block handling into one,
which can incur unnecessary overhead (with preprocessing the files and
data storage) and makes the semantics messy, we separated them into
BlockReader, which can only read parts of the block and get the size of
it, BlockAppender, which can only create a new block and execute
append-only writes to it, and BlockOperator, which can only execute
whole-block operations like copy, move, delete, etc.

We then eliminated the BlockHandler and BlockHandlerLocal classes and
replaced all usages with the appropriate specialized handler.