Removed passing -1 as length argument when reading
When we have partial block caching, BlockReaders won't always have the
entire block in the directory they're reading. Thus when we're reading a
block with a BlockReader, the reader can't always know the total length
of the block. Then for the BlockReader, reading till the end of the
block doesn't make sense, because we don't know where the end of the
block is. Thus passing a -1 pretty much anywhere we might end up reading
from a BlockReader doesn't make sense anymore, so we eliminate that
option.