Sped up BlockInStream remote read
The remote read was refreshing the ClientBlockInfo every time it failed
a remote read, which is pretty expensive. The current Tachyon solves
this by disabling remote reads when we fail once, but this might not
work so well for partial caching, since reads may fail due to the
workers not having a certain part of the block, while they may have
another part of the block.

So instead, we don't refresh ClientBlockInfo and only try the workers
once each.