Fixed remoteReadByteBuffer spec in TachyonFile
Since the new BlockInStream will NOT check the local worker in the
readRemoteByteBuffer method, the TachyonFile#remoteReadByteBuffer method
will not check the local worker.

This is not a problem in practice, because TachyonFile will always try a
local read before going to the remote workers, but we needed to amend
the unit test for TachyonFile.