Compromised between tachyon-perf and conviva for read strategy
Based on some rough benchmarks, it seems like reading 1-byte at a
time (or generally small chunks of bytes at a time) is better to do from
a MappedByteBuffer than directly from the FileChannel, but then reading
large chunks is better done directly from the FileChannel.

So in BlockInStream, we guess that if the read covers the entire
remainder of the channel, we read directly from the FileChannel, but
otherwise we map the channel. This seems to get a good compromise
between tachyon-perf performance and conviva performance for now.