Improved BlockInStream reading
Based on some measurements, reading from a FileChannel is better for
large-enough reads, and from a ByteBuffer is better for smaller reads.
Thus the BlockInStream now keeps track of the average read size, and
picks to read directly from the FileChannel or to map the ByteBuffer
based on whether this average exceeds the experimental threshold.